CLOUDP-233673 Move Atlas Deployments tests
