+++ tools/genevergreen/main.go
@@ -66,8 +66,6 @@ func run() error {
 		generate.PublishSnapshotTasks(c, toolName)
 	case "publish":
 		generate.PublishStableTasks(c, toolName)
-	case "local":
-		generate.LocalDeploymentTasks(c, toolName)
 	default:
 		return errors.New("-tasks is invalid")
 	}
